Analyse drug set enrichment

Analyse drug set enrichment

analyseDrugSetEnrichment(
  sets,
  stats,
  col = NULL,
  nperm = 10000,
  maxSize = 500,
  ...
)

Arguments

sets
Named list of characters: named sets containing compound identifiers (obtain drug sets by running prepareDrugSets())

stats
Named numeric vector or either a similarPerturbations or a targetingDrugs object (obtained after running rankSimilarPerturbations or predictTargetingDrugs, respectively)

col
Character: name of the column to use for statistics (only required if class of stats is either similarPerturbations or targetingDrugs)

nperm
Number of permutations to do. Minimial possible nominal p-value is about 1/nperm

maxSize
Maximal size of a gene set to test. All pathways above the threshold are excluded.

...
Arguments passed on to fgsea::fgsea

minSize

Minimal size of a gene set to test. All pathways below the threshold are excluded.

nproc

If not equal to zero sets BPPARAM to use nproc workers (default = 0).

gseaParam

GSEA parameter value, all gene-level statis are raised to the power of `gseaParam` before calculation of GSEA enrichment scores.

BPPARAM

Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting `nproc` default value `bpparam()` is used.

Value

Enrichment analysis based on GSEA

See also

Examples

descriptors <- loadDrugDescriptors()

#> molecular_descriptors_NCI60_2D.rds not found: downloading data...

drugSets <- prepareDrugSets(descriptors)

# Analyse drug set enrichment in ranked targeting drugs for a differential
# expression profile
data("diffExprStat")
gdsc      <- loadExpressionDrugSensitivityAssociation("GDSC")

#> expressionDrugSensitivityCorGDSC_7.rds not found: downloading data...

#> Loading data from expressionDrugSensitivityCorGDSC_7.rds...

predicted <- predictTargetingDrugs(diffExprStat, gdsc)

#> Subsetting data based on 11396 intersecting genes (85% of the 13451 input genes)...

#> Correlating against 266 GDSC 7 compounds (983 cell lines; Spearman's correlation)...

#> Comparison against 266 GDSC 7 compounds using 'spearman' performed in 2.15 secs

#> Correlating against 266 GDSC 7 compounds (983 cell lines; Pearson's correlation)...

#> Comparison against 266 GDSC 7 compounds using 'pearson' performed in 0.68 secs

#> Performing GSEA against 266 GDSC 7 compounds (983 cell lines)......

#> Comparison against 266 GDSC 7 compounds using 'gsea' (gene size of 150) performed in 5.83 secs

analyseDrugSetEnrichment(drugSets, predicted)

#> Matching compounds with those available in drug sets...

#> Performing enrichment analysis...

#>                                     pathway        pval       padj         ES
#>  1:              FDA_status: Clinical trial 0.003254770 0.09562058 -0.7395954
#>  2:            Reproductive Effective: high 0.005428292 0.09562058 -0.6651727
#>  3:                   Aromatic Nitrogens: 6 0.005737235 0.09562058  0.9571429
#>  4:                         Mutagenic: high 0.016049383 0.11463845 -0.6708012
#>  5:                          Irritant: high 0.013559673 0.11463845 -0.5684080
#>  6:                               Amines: 0 0.014109877 0.11463845 -0.4484031
#>  7:                   Aromatic Nitrogens: 2 0.011547833 0.11463845 -0.6108034
#>  8:                         Alkyl-Amines: 0 0.020618557 0.12886598 -0.4284566
#>  9:                               Amines: 1 0.031386491 0.16254876 -0.5018454
#> 10:                      Aromatic Amines: 0 0.032509753 0.16254876 -0.4013689
#> 11:                       Aromatic Rings: 5 0.039541161 0.17110710 -0.7921147
#> 12:                       Aromatic Rings: 6 0.041065705 0.17110710 -0.9822064
#> 13:                       Acidic Oxygens: 0 0.044904490 0.17270958 -0.3841362
#> 14:                       Tumorigenic: high 0.062348267 0.19472535 -0.5814620
#> 15:                       Tumorigenic: none 0.060606061 0.19472535 -0.3818599
#> 16:                          Metal-Atoms: 0 0.066206621 0.19472535 -0.3715652
#> 17:                      Basic Nitrogens: 0 0.065685391 0.19472535 -0.4153448
#> 18:                         Mutagenic: none 0.079607961 0.19664647 -0.3729461
#> 19:            Reproductive Effective: none 0.107810781 0.19664647 -0.3642183
#> 20:                          Irritant: none 0.110122024 0.19664647 -0.3675231
#> 21:                            Fragments: 1 0.082716543 0.19664647 -0.3772780
#> 22:                            Fragments: 2 0.091049542 0.19664647 -0.4775643
#> 23:                       Aromatic Rings: 0 0.108170552 0.19664647 -0.4935085
#> 24:                   Aromatic Nitrogens: 4 0.071445668 0.19664647 -0.6558464
#> 25:                      Basic Nitrogens: 2 0.099586821 0.19664647 -0.5106646
#> 26:                      Basic Nitrogens: 4 0.097804391 0.19664647  0.9537367
#> 27:          Stereo Configuration: racemate 0.099372225 0.19664647 -0.8214286
#> 28:   Stereo Configuration: this enantiomer 0.097206704 0.19664647 -0.4506119
#> 29:                   Aromatic Nitrogens: 0 0.127420488 0.21236748 -0.3964133
#> 30:                       Acidic Oxygens: 1 0.123717654 0.21236748 -0.7998313
#> 31:                   Aromatic Nitrogens: 3 0.138048996 0.22265967 -0.5308021
#> 32:                          Metal-Atoms: 1 0.144030449 0.22504758 -0.9288256
#> 33:                       Aromatic Rings: 4 0.162516929 0.24623777 -0.4578404
#> 34:                FDA_status: FDA approved 0.187456237 0.26779462 -0.3478354
#> 35: Stereo Configuration: unknown chirality 0.185909980 0.26779462 -0.4363479
#> 36:                       Aromatic Rings: 3 0.255024745 0.34895461 -0.3836849
#> 37:                               Amines: 3 0.272184594 0.34895461 -0.5108696
#> 38:                      Aromatic Amines: 1 0.260860506 0.34895461 -0.4258960
#> 39:                      Basic Nitrogens: 3 0.271151418 0.34895461 -0.5113624
#> 40:                         Alkyl-Amines: 1 0.279840033 0.34980004 -0.3999088
#> 41:                               Amines: 2 0.333591331 0.40681870  0.4202899
#> 42:                   Aromatic Nitrogens: 5 0.428685897 0.51034035 -0.7864769
#> 43:                      Basic Nitrogens: 1 0.479539902 0.55760454 -0.3478278
#> 44:                       Acidic Oxygens: 2 0.497205589 0.56500635  0.7544484
#> 45:                          Mutagenic: low 0.533333333 0.57905337  0.7330961
#> 46:                        Tumorigenic: low 0.533333333 0.57905337  0.7330961
#> 47:                       Aromatic Rings: 2 0.554434030 0.57905337 -0.3405766
#> 48:                   Aromatic Nitrogens: 1 0.555891239 0.57905337 -0.3506263
#> 49:                       Aromatic Rings: 1 0.623408126 0.63613074  0.3682310
#> 50:                         Alkyl-Amines: 2 0.789378072 0.78937807 -0.3047590
#>                                     pathway        pval       padj         ES
#>            NES nMoreExtreme size                                   leadingEdge
#>  1: -1.7430029           28    7                   174939,656576,330507,706995
#>  2: -1.6637438           49    9         174939,287459,63878,759877,732517,185
#>  3:  1.5868793           19    2                                 125066,758612
#>  4: -1.5808759          142    7                    752782,737754,287459,63878
#>  5: -1.5696148          130   14      752782,737754,287459,63878,759877,732517
#>  6: -1.4531939          140   40 174939,656576,752782,737754,759850,628503,...
#>  7: -1.6029012          108   11              656576,745750,727989,681239,1014
#>  8: -1.4034833          205   45 174939,656576,752782,737754,759850,628503,...
#>  9: -1.4657763          308   19       706995,745750,727989,1011,759877,732517
#> 10: -1.3427976          324   58 174939,656576,752782,737754,759850,628503,...
#> 11: -1.4489833          292    3                          745750,727989,747599
#> 12: -1.3102354          204    1                                        656576
#> 13: -1.3010198          448   68 174939,656576,752782,737754,759850,628503,...
#> 14: -1.4126881          564    8                        1011,287459,63878,1014
#> 15: -1.2842916          605   62 174939,656576,752782,737754,759850,628503,...
#> 16: -1.2610882          661   70 174939,656576,752782,737754,759850,628503,...
#> 17: -1.3255085          655   35 174939,656576,759850,628503,119875,330507,...
#> 18: -1.2562423          795   63 174939,656576,759850,628503,119875,330507,...
#> 19: -1.2249584         1077   62 656576,752782,737754,759850,628503,119875,...
#> 20: -1.2280085         1100   57 174939,656576,759850,628503,119875,330507,...
#> 21: -1.2586467          826   56 174939,656576,752782,759850,628503,119875,...
#> 22: -1.3377566          883   15        737754,727989,63878,757384,90636,49842
#> 23: -1.3196241         1029   12     119875,330507,287459,63878,683864,185,...
#> 24: -1.4039553          596    5                                 752782,737754
#> 25: -1.3401119          939   11              752782,737754,757384,90636,49842
#> 26:  1.2691779          489    1                                        758612
#> 27: -1.3171340          648    2                                     1011,1014
#> 28: -1.3161350          956   19   656576,628503,330507,681239,90636,49842,...
#> 29: -1.2458887         1269   31 174939,759850,628503,119875,330507,287459,...
#> 30: -1.2825034          807    2                                          1011
#> 31: -1.2896075         1250    8                          757441,759877,732517
#> 32: -1.2390270          718    1                                        119875
#> 33: -1.2438086         1559   13     752782,737754,757441,1011,1014,758487,...
#> 34: -1.1636988         1873   58 752782,737754,759850,628503,119875,745750,...
#> 35: -1.2223010         1804   15  759850,287459,63878,757384,758645,757804,...
#> 36: -1.1490826         2524   22   706995,759877,732517,757384,90636,49842,...
#> 37: -1.1557393         2370    6                            757384,90636,49842
#> 38: -1.1570257         2503   13                706995,1011,757384,90636,49842
#> 39: -1.1568544         2361    6                                 759877,732517
#> 40: -1.1341868         2728   16                   745750,727989,759877,732517
#> 41:  1.0982605          430    6     125066,758612,758247,743414,764040,758487
#> 42: -1.0491378         2139    1                                        747599
#> 43: -1.0068789         4710   18               745750,727989,287459,63878,1014
#> 44:  1.0039765         2490    1                                           740
#> 45:  0.9755621         2671    1                                            17
#> 46:  0.9755621         2671    1                                            17
#> 47: -0.9540257         5382   15 174939,759850,628503,681239,758246,757804,...
#> 48: -0.9525417         5335   13     706995,1011,757384,90636,49842,684363,...
#> 49:  0.8874347         1027    5                   295,17,701852,759852,684363
#> 50: -0.7835660         7386   10                            757384,90636,49842
#>            NES nMoreExtreme size                                   leadingEdge