Download metadata for ENCODE knockdown experiments

Download metadata for ENCODE knockdown experiments

downloadENCODEknockdownMetadata(
  cellLine = NULL,
  gene = NULL,
  file = "ENCODEmetadata.rds"
)

Arguments

cellLine
Character: cell line

gene
Character: target gene

file
Character: RDS file with metadata (if file doesn't exist, it will be created)

Value

Data frame containing ENCODE knockdown experiment metadata

See also

Examples

downloadENCODEknockdownMetadata("HepG2", "EIF4G1")

#>    File accession File format File type File format type          Output type
#> 1:    ENCFF955TXI         tsv       tsv                  gene quantifications
#> 2:    ENCFF049UZV         tsv       tsv                  gene quantifications
#>    File assembly Experiment accession         Assay Biosample term id
#> 1:          hg19          ENCSR509LIV shRNA RNA-seq       EFO:0001187
#> 2:          hg19          ENCSR509LIV shRNA RNA-seq       EFO:0001187
#>    Biosample term name Biosample type Biosample organism Biosample treatments
#> 1:               HepG2      cell line       Homo sapiens                   NA
#> 2:               HepG2      cell line       Homo sapiens                   NA
#>    Biosample treatments amount Biosample treatments duration
#> 1:                          NA                            NA
#> 2:                          NA                            NA
#>    Biosample genetic modifications methods
#> 1:                                    RNAi
#> 2:                                    RNAi
#>    Biosample genetic modifications categories
#> 1:                               interference
#> 2:                               interference
#>    Biosample genetic modifications targets
#> 1:                  /targets/EIF4G1-human/
#> 2:                  /targets/EIF4G1-human/
#>    Biosample genetic modifications gene targets
#> 1:                                           NA
#> 2:                                           NA
#>    Biosample genetic modifications site coordinates
#> 1:                                               NA
#> 2:                                               NA
#>    Biosample genetic modifications zygosity Experiment target
#> 1:                                       NA            EIF4G1
#> 2:                                       NA            EIF4G1
#>      Library made from Library depleted in
#> 1: polyadenylated mRNA                    
#> 2: polyadenylated mRNA                    
#>                                    Library extraction method
#> 1: Maxwell 16 LEV simpleRNA Cells Kit (Promega cat#: AS1270)
#> 2: Maxwell 16 LEV simpleRNA Cells Kit (Promega cat#: AS1270)
#>    Library lysis method Library crosslinking method Library strand specific
#> 1:                   NA                          NA         strand-specific
#> 2:                   NA                          NA         strand-specific
#>    Experiment date released Project RBNS protein concentration
#> 1:               2015-08-18  ENCODE                         NA
#> 2:               2015-08-18  ENCODE                         NA
#>    Library fragmentation method Library size range Biological replicate(s)
#> 1:   chemical (Illumina TruSeq)               >200                       1
#> 2:   chemical (Illumina TruSeq)               >200                       2
#>    Technical replicate(s) Read length Mapped read length Run type Paired end
#> 1:                    1_1          NA                 NA                  NA
#> 2:                    2_1          NA                 NA                  NA
#>    Paired with Index of                             Derived from    Size
#> 1:                   NA /files/ENCFF995LZN/, /files/ENCFF826ONU/ 9554701
#> 2:                   NA /files/ENCFF447GOE/, /files/ENCFF826ONU/ 9563218
#>                           Lab                           md5sum dbxrefs
#> 1: ENCODE Processing Pipeline f9b46cbff89bfe91bdb0879b81cf41e8        
#> 2: ENCODE Processing Pipeline 9f77694ffdf48801edb9c39f75f6d3b7        
#>                                                             File download URL
#> 1: https://www.encodeproject.org/files/ENCFF955TXI/@@download/ENCFF955TXI.tsv
#> 2: https://www.encodeproject.org/files/ENCFF049UZV/@@download/ENCFF049UZV.tsv
#>    Genome annotation Platform Controlled by File Status
#> 1:               V19                           released
#> 2:               V19                           released
#>                                                                                s3_uri
#> 1: s3://encode-public/2015/08/26/f02dd75d-a0c9-451f-aae8-23eb5e8a5a2f/ENCFF955TXI.tsv
#> 2: s3://encode-public/2015/08/26/d03e3680-479e-470b-9126-198e6cefe4dc/ENCFF049UZV.tsv
#>                         Audit WARNING Audit NOT_COMPLIANT Audit ERROR
#> 1: missing biosample characterization                                
#> 2: missing biosample characterization                                
#>        Control      Control Experiments
#> 1: ENCSR042QTH ENCFF657KMW, ENCFF726AVT
#> 2: ENCSR042QTH ENCFF657KMW, ENCFF726AVT