Load table with drug descriptors

Load table with drug descriptors

loadDrugDescriptors(
  source = c("NCI60", "CMap"),
  type = c("2D", "3D"),
  file = NULL,
  path = NULL
)

Arguments

source
Character: source of compounds used to calculate molecular descriptors (NCI60 or CMap)

type
Character: load 2D or 3D molecular descriptors

file
Character: filepath to drug descriptors (automatically downloaded if file does not exist)

path
Character: folder where to find files (optional; file may contain the full filepath if preferred)

Value

Data table with drug descriptors

See also

Examples

loadDrugDescriptors()
#>        compound FDA_status Mechanism_of_action Total Molweight
#>     1:        1       <NA>                <NA>         122.123
#>     2:       17       <NA>                <NA>         319.531
#>     3:       89       <NA>                <NA>         212.699
#>     4:      185       <NA>                <NA>         281.351
#>     5:      295       <NA>                <NA>         164.203
#>    ---                                                        
#> 21734:   780184       <NA>                <NA>         485.555
#> 21735:   780185       <NA>                <NA>         520.000
#> 21736:   780186       <NA>                <NA>         545.607
#> 21737:   780198       <NA>                <NA>         474.507
#> 21738:   783538       <NA>                <NA>         654.663
#>        Monoisotopic Mass  cLogP  cLogS H-Acceptors H-Donors Total Surface Area
#>     1:          122.0368 0.5246 -1.122           2        0              98.97
#>     2:          319.2875 7.3033 -5.409           2        2             293.03
#>     3:          177.1154 1.2787 -1.693           2        0             153.35
#>     4:          281.1627 0.9427 -2.688           5        2             213.38
#>     5:          164.0837 2.0516 -2.134           2        1             138.64
#>    ---                                                                        
#> 21734:          485.1508 3.8839 -5.205           8        2             366.97
#> 21735:          519.1119 4.4217 -5.911           8        1             386.04
#> 21736:          545.1720 3.6757 -5.211          10        1             415.14
#> 21737:          474.1679 4.3868 -6.633           7        0             369.08
#> 21738:          654.2312 2.2501 -4.057          13        4             460.31
#>        Polar Surface Area Druglikeness Mutagenic Tumorigenic
#>     1:              34.14     -0.69205      high        none
#>     2:              46.25    -19.80400       low         low
#>     3:              20.31      3.55190      none        none
#>     4:              83.47      4.15230      none        none
#>     5:              37.30     -6.26530      none        none
#>    ---                                                      
#> 21734:             111.70     -0.58316      none        none
#> 21735:             100.70      1.25170      none        none
#> 21736:             119.16      5.34740      none        none
#> 21737:              72.45     -2.40640      high        none
#> 21738:             171.83     -4.11790      none        none
#>        Reproductive Effective Irritant Nasty Functions Shape Index
#>     1:                   none     none            <NA>     0.66667
#>     2:                   none     none            <NA>     0.82609
#>     3:                   none     none            <NA>     0.69231
#>     4:                   high     high            <NA>     0.50000
#>     5:                   none     none            <NA>     0.75000
#>    ---                                                            
#> 21734:                   high     none            <NA>     0.50000
#> 21735:                   high     none            <NA>     0.48571
#> 21736:                   high     none            <NA>     0.50000
#> 21737:                   high     none            <NA>     0.51429
#> 21738:                   high     none            <NA>     0.38298
#>        Molecular Flexibility Molecular Complexity Fragments Non-H Atoms
#>     1:               0.24673              0.72259         1           9
#>     2:               0.48864              0.60279         1          23
#>     3:               0.55528              0.52258         2          13
#>     4:               0.49493              0.70898         1          20
#>     5:               0.57069              0.43984         1          12
#>    ---                                                                 
#> 21734:               0.48741              0.86255         1          34
#> 21735:               0.48769              0.87147         1          35
#> 21736:               0.50651              0.88979         1          38
#> 21737:               0.46227              0.86525         1          35
#> 21738:               0.29662              1.08780         1          47
#>        Non-C/H Atoms Metal-Atoms Electronegative Atoms Stereo Centers
#>     1:             2           0                     2              0
#>     2:             2           0                     2              0
#>     3:             2           0                     2              0
#>     4:             5           0                     5              4
#>     5:             2           0                     2              0
#>    ---                                                               
#> 21734:             9           0                     9              0
#> 21735:            10           0                    10              0
#> 21736:            11           0                    11              0
#> 21737:             7           0                     7              0
#> 21738:            13           0                    13              9
#>        Rotatable Bonds Rings Closures Small Rings Aromatic Rings Aromatic Atoms
#>     1:               0              1           1              0              0
#>     2:              14              1           1              1              6
#>     3:               4              1           1              1              6
#>     4:               3              2           2              0              0
#>     5:               4              1           1              1              6
#>    ---                                                                         
#> 21734:               8              3           3              3             18
#> 21735:               9              3           3              3             18
#> 21736:              11              3           3              3             18
#> 21737:               9              4           4              3             18
#> 21738:              11              6           6              3             18
#>        sp3-Atoms Symmetric atoms Amides Amines Alkyl-Amines Aromatic Amines
#>     1:         1               0      0      0            0               0
#>     2:        16               0      0      1            0               1
#>     3:         5               3      0      1            1               0
#>     4:        13               3      1      0            0               0
#>     5:         4               2      0      0            0               0
#>    ---                                                                     
#> 21734:        11               7      1      0            0               0
#> 21735:        11               5      1      0            0               0
#> 21736:        15               5      1      0            0               0
#> 21737:        11               4      0      0            0               0
#> 21738:        27               4      0      0            0               0
#>        Aromatic Nitrogens Basic Nitrogens Acidic Oxygens Stereo Configuration
#>     1:                  0               0              0                 <NA>
#>     2:                  0               0              0                 <NA>
#>     3:                  0               1              0                 <NA>
#>     4:                  0               0              0      this enantiomer
#>     5:                  0               0              1                 <NA>
#>    ---                                                                       
#> 21734:                  0               0              0                 <NA>
#> 21735:                  0               0              0                 <NA>
#> 21736:                  0               0              0                 <NA>
#> 21737:                  0               0              0                 <NA>
#> 21738:                  0               0              0      this enantiomer