Plot drug set enrichment

Plot drug set enrichment

plotDrugSetEnrichment(
  sets,
  stats,
  col = "rankProduct_rank",
  selectedSets = NULL,
  keyColSets = NULL,
  keyColStats = NULL
)

Arguments

sets

Named list of characters: named sets containing compound identifiers (obtain drug sets by running prepareDrugSets())

stats

Named numeric vector or either a similarPerturbations or a targetingDrugs object (obtained after running rankSimilarPerturbations or predictTargetingDrugs, respectively)

col

Character: name of the column to use for statistics (only required if class of stats is either similarPerturbations or targetingDrugs)

selectedSets

Character: drug sets to plot (if NULL, plot all)

keyColSets

Character: column from sets to compare with column keyColStats from stats; automatically selected if NULL

keyColStats

Character: column from stats to compare with column keyColSets from sets; automatically selected if NULL

Value

List of GSEA plots per drug set

See also

Examples

descriptors <- loadDrugDescriptors()
drugSets <- prepareDrugSets(descriptors)

# Analyse drug set enrichment in ranked targeting drugs for a differential
# expression profile
data("diffExprStat")
gdsc      <- loadExpressionDrugSensitivityAssociation("GDSC")
#> Loading data from expressionDrugSensitivityCorGDSC7.qs...
#> Error in qread(file): something went wrong (reading object header)
predicted <- predictTargetingDrugs(diffExprStat, gdsc)
#> Error in eval(expr, envir, enclos): object 'gdsc' not found

plotDrugSetEnrichment(drugSets, predicted)
#> Matching compounds with those available in drug sets...
#> Error in eval(expr, envir, enclos): object 'predicted' not found