Predict targeting drugs

Identify compounds that may target the phenotype associated with a user-provided differential expression profile by comparing such against a correlation matrix of gene expression and drug sensitivity.

predictTargetingDrugs(
  input,
  expressionDrugSensitivityCor,
  method = c("spearman", "pearson", "gsea"),
  geneSize = 150,
  isDrugActivityDirectlyProportionalToSensitivity = NULL
)

Arguments

input
Named numeric vector of differentially expressed genes whose names are gene identifiers and respective values are a statistic that represents significance and magnitude of differentially expressed genes (e.g. t-statistics); or character of gene symbols composing a gene set that is tested for enrichment in reference data (only used if method includes gsea)

expressionDrugSensitivityCor
Matrix: correlation matrix of gene expression (rows) and drug sensitivity (columns) across cell lines. Pre-prepared gene expression and drug sensitivity associations are available to download using loadExpressionDrugSensitivityAssociation().

method
Character: one or more methods to compare data (spearman, pearson or gsea)

geneSize
Numeric: number of top up-/down-regulated genes to use as gene sets to test for enrichment in reference data; if a 2-length numeric vector, the first index is the number of top up-regulated genes and the second index is the number of down-regulated genes used to create gene sets; only used if method includes gsea and if input is not a gene set

isDrugActivityDirectlyProportionalToSensitivity
Boolean: are the values used for drug activity directly proportional to drug sensitivity? See details.

Value

Data table with correlation or GSEA results comparing differential expression values against gene expression and drug sensitivity associations

Details

If isDrugActivityDirectlyProportionalToSensitivity = NULL, the attribute isDrugMetricDirectlyProportionalToSensitivity of expressionDrugSensitivityCor is used (see loadExpressionDrugSensitivityAssociation()).

GSEA score

Weighted connectivity scores (WTCS) are calculated when method = "gsea" (https://clue.io/connectopedia/cmap_algorithms).

See also

Examples

# Example of a differential expression profile
data("diffExprStat")

# Load expression and drug sensitivity association derived from GDSC data
gdsc <- loadExpressionDrugSensitivityAssociation("GDSC 7")

#> Loading data from expressionDrugSensitivityCorGDSC_7.rds...

# Predict targeting drugs on a differential expression profile
predictTargetingDrugs(diffExprStat, gdsc)

#> Subsetting data based on 11396 intersecting genes (85% of the 13451 input genes)...

#> Correlating against 266 GDSC 7 compounds (983 cell lines; Spearman's correlation)...

#> Comparison against 266 GDSC 7 compounds using 'spearman' performed in 2.48 secs

#> Correlating against 266 GDSC 7 compounds (983 cell lines; Pearson's correlation)...

#> Comparison against 266 GDSC 7 compounds using 'pearson' performed in 0.8 secs

#> Performing GSEA against 266 GDSC 7 compounds (983 cell lines)......

#> Comparison against 266 GDSC 7 compounds using 'gsea' (gene size of 150) performed in 13.18 secs

#>      compound spearman_coef spearman_pvalue spearman_qvalue pearson_coef
#>   1:     1047    0.14532331    7.908329e-55    2.103616e-52   0.13098933
#>   2:      207    0.10447948    4.950795e-29    2.633823e-27   0.09295384
#>   3:      157    0.12935630    1.020442e-43    1.357188e-41   0.11655136
#>   4:     1091    0.08481582    1.194387e-19    1.512890e-18   0.08245367
#>   5:      110    0.11857369    5.806289e-37    5.148243e-35   0.11346259
#>  ---                                                                    
#> 262:     1133   -0.02817757    2.627306e-03    3.529613e-03  -0.02761751
#> 263:     1149   -0.01812593    5.299930e-02    6.210491e-02  -0.02690099
#> 264:       35   -0.02099097    2.503708e-02    3.069061e-02  -0.03075034
#> 265:     1029   -0.04152497    9.237123e-06    1.565016e-05  -0.04525941
#> 266:     1031   -0.08489029    1.109885e-19    1.476147e-18  -0.08808410
#>      pearson_pvalue pearson_qvalue       GSEA spearman_rank pearson_rank
#>   1:   8.577544e-45   2.281627e-42  0.0000000             1            1
#>   2:   2.693140e-23   1.023393e-21  0.2769544             5            7
#>   3:   9.209531e-36   1.224868e-33  0.0000000             2            2
#>   4:   1.184660e-18   2.100798e-17  0.3118949            20           14
#>   5:   5.716381e-34   5.068524e-32  0.0000000             3            3
#>  ---                                                                    
#> 262:   3.193570e-03   4.591836e-03 -0.2605952           263          261
#> 263:   4.079592e-03   5.772189e-03 -0.3444514           259          260
#> 264:   1.026803e-03   1.587963e-03 -0.3038956           261          263
#> 265:   1.341660e-06   2.949435e-06 -0.3298198           264          264
#> 266:   4.491968e-21   1.194863e-19 -0.3800912           266          266
#>      GSEA_rank rankProduct_rank
#>   1:     130.5                1
#>   2:       9.0                2
#>   3:     130.5                3
#>   4:       3.0                4
#>   5:     130.5                5
#>  ---                           
#> 262:     257.0              262
#> 263:     265.0              263
#> 264:     262.0              264
#> 265:     264.0              265
#> 266:     266.0              266