Prepare CMap perturbation data

Prepare CMap perturbation data

prepareCMapPerturbations(
  metadata,
  zscores,
  geneInfo,
  compoundInfo = NULL,
  ...,
  loadZscores = FALSE
)

Arguments

metadata

Data frame (CMap metadata) or character (respective filepath to load data from file)

zscores

Data frame (GCTX z-scores) or character (respective filepath to load data from file)

geneInfo

Data frame (CMap gene info) or character (respective filepath to load data from file)

compoundInfo

Data frame (CMap compound info) or character (respective filepath to load data from file)

...

Arguments passed on to filterCMapMetadata

cellLine

Character: cell line (if NULL, all values are loaded)

timepoint

Character: timepoint (if NULL, all values are loaded)

dosage

Character: dosage (if NULL, all values are loaded)

perturbationType

Character: type of perturbation (if NULL, all perturbation types are loaded)

loadZscores

Boolean: load matrix of perturbation z-scores? Not recommended in systems with less than 30GB of RAM; if FALSE, downstream functions will load and process the file directly chunk by chunk, resulting in a lower memory footprint

Value

CMap perturbation data attributes and filename

See also

Examples

metadata <- loadCMapData("cmapMetadata.txt", "metadata")
#> Loading CMap metadata from cmapMetadata.txt...
metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
if (FALSE) {
prepareCMapPerturbations(metadata, "cmapZscores.gctx", "cmapGeneInfo.txt")
}