Analyse drug set enrichment

Analyse drug set enrichment

analyseDrugSetEnrichment(
  sets,
  stats,
  col = NULL,
  nperm = 10000,
  maxSize = 500,
  ...
)

Arguments

sets
Named list of characters: named sets containing compound identifiers (obtain drug sets by running prepareDrugSets())

stats
Named numeric vector or either a similarPerturbations or a targetingDrugs object (obtained after running rankSimilarPerturbations or predictTargetingDrugs, respectively)

col
Character: name of the column to use for statistics (only required if class of stats is either similarPerturbations or targetingDrugs)

nperm
Number of permutations to do. Minimial possible nominal p-value is about 1/nperm

maxSize
Maximal size of a gene set to test. All pathways above the threshold are excluded.

...
Arguments passed on to fgsea::fgsea

minSize

Minimal size of a gene set to test. All pathways below the threshold are excluded.

nproc

If not equal to zero sets BPPARAM to use nproc workers (default = 0).

gseaParam

GSEA parameter value, all gene-level statis are raised to the power of `gseaParam` before calculation of GSEA enrichment scores.

BPPARAM

Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting `nproc` default value `bpparam()` is used.

Value

Enrichment analysis based on GSEA

See also

Examples

descriptors <- loadDrugDescriptors()

#> molecular_descriptors_NCI60_2D.rds not found: downloading data...

drugSets <- prepareDrugSets(descriptors)

# Analyse drug set enrichment in ranked targeting drugs for a differential
# expression profile
data("diffExprStat")
gdsc      <- loadExpressionDrugSensitivityAssociation("GDSC")

#> expressionDrugSensitivityCorGDSC_7.rds not found: downloading data...

#> Loading data from expressionDrugSensitivityCorGDSC_7.rds...

predicted <- predictTargetingDrugs(diffExprStat, gdsc)

#> Subsetting data based on 11396 intersecting genes (85% of the 13451 input genes)...

#> Correlating against 266 GDSC 7 compounds (983 cell lines; Spearman's correlation)...

#> Comparison against 266 GDSC 7 compounds using 'spearman' performed in 2.33 secs

#> Correlating against 266 GDSC 7 compounds (983 cell lines; Pearson's correlation)...

#> Comparison against 266 GDSC 7 compounds using 'pearson' performed in 0.72 secs

#> Performing GSEA against 266 GDSC 7 compounds (983 cell lines)......

#> Comparison against 266 GDSC 7 compounds using 'gsea' (gene size of 150) performed in 8.97 secs

analyseDrugSetEnrichment(drugSets, predicted)

#> Matching compounds with those available in drug sets...

#> Performing enrichment analysis...

#>                             pathway         pval       padj         ES
#>   1:                 Small Rings: 4 0.0002091175 0.02969469 -0.7423058
#>   2:                   sp3-Atoms: 5 0.0001191469 0.02969469 -0.9039180
#>   3:              H-Donors: [6, 40] 0.0014409222 0.13640730  0.8884892
#>   4: Polar Surface Area: [123, 141] 0.0019991582 0.14194024 -0.6515166
#>   5:     FDA_status: Clinical trial 0.0044147612 0.18258633 -0.7395954
#>  ---                                                                  
#> 280:                    H-Donors: 1 0.9607488565 0.97447384 -0.2272251
#> 281:    Total Molweight: [294, 312] 0.9853853854 0.98886731 -0.5088968
#> 282:              Stereo Centers: 3 0.9853853854 0.98886731 -0.5088968
#> 283:                 Small Rings: 3 0.9804475115 0.98886731 -0.1975991
#> 284:              Rings Closures: 3 0.9970002143 0.99700021 -0.1762806
#>             NES nMoreExtreme size                                   leadingEdge
#>   1: -2.0146498            1   13   752782,737754,757441,1011,759877,732517,...
#>   2: -1.9454159            0    5                          174939,752782,737754
#>   3:  1.9835693            2    4                   125066,758612,123127,759155
#>   4: -1.7401222           18   12  174939,752782,737754,287459,63878,759877,...
#>   5: -1.7393392           38    7                   174939,656576,330507,706995
#>  ---                                                                           
#> 280: -0.5955060         9031   11            656576,758487,764040,753686,747856
#> 281: -0.6776878         4921    1                                        613327
#> 282: -0.6776878         4921    1                                        613327
#> 283: -0.5362925         9376   13 706995,683864,339555,758664,606698,226080,...
#> 284: -0.4531435         9305   10                                        706995