Analyse drug set enrichment

Analyse drug set enrichment

analyseDrugSetEnrichment(
  sets,
  stats,
  col = NULL,
  nperm = 10000,
  maxSize = 500,
  ...,
  keyColSets = NULL,
  keyColStats = NULL
)

Arguments

sets
Named list of characters: named sets containing compound identifiers (obtain drug sets by running prepareDrugSets())

stats
Named numeric vector or either a similarPerturbations or a targetingDrugs object (obtained after running rankSimilarPerturbations or predictTargetingDrugs, respectively)

col
Character: name of the column to use for statistics (only required if class of stats is either similarPerturbations or targetingDrugs)

nperm
Number of permutations to do. Minimial possible nominal p-value is about 1/nperm

maxSize
Maximal size of a gene set to test. All pathways above the threshold are excluded.

...
Arguments passed on to fgsea::fgseaSimple

minSize

Minimal size of a gene set to test. All pathways below the threshold are excluded.

scoreType

This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")

nproc

If not equal to zero sets BPPARAM to use nproc workers (default = 0).

gseaParam

GSEA parameter value, all gene-level statis are raised to the power of `gseaParam` before calculation of GSEA enrichment scores.

BPPARAM

Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting `nproc` default value `bpparam()` is used.

keyColSets
Character: column from sets to compare with column keyColStats from stats; automatically selected if NULL

keyColStats
Character: column from stats to compare with column keyColSets from sets; automatically selected if NULL

Value

Enrichment analysis based on GSEA

See also

Examples

descriptors <- loadDrugDescriptors()

#> molecular_descriptors_NCI60_2D.rds not found: downloading data...

drugSets <- prepareDrugSets(descriptors)

# Analyse drug set enrichment in ranked targeting drugs for a differential
# expression profile
data("diffExprStat")
gdsc      <- loadExpressionDrugSensitivityAssociation("GDSC")

#> expressionDrugSensitivityCorGDSC7.rds not found: downloading data...

#> Loading data from expressionDrugSensitivityCorGDSC7.rds...

predicted <- predictTargetingDrugs(diffExprStat, gdsc)

#> Subsetting data based on 11396 intersecting genes (85% of the 13451 input genes)...

#> Comparing against 266 GDSC 7 compounds (983 cell lines) using 'spearman, pearson, gsea' (gene size of 150)...

#> Comparison performed in 3.84 secs

analyseDrugSetEnrichment(drugSets, predicted)

#> Matching compounds with those available in drug sets...

#> Ordering results by column 'rankProduct_rank'; to manually select column to order by, please set argument 'col'

#> Columns 'name' and 'name' were matched based on 52 common values; to manually select columns to compare, please set arguments starting with 'keyCol'

#> Performing enrichment analysis...

#>                       descriptor         pval       padj         ES        NES
#>   1:              Small Rings: 4 0.0001043623 0.02963891 -0.7423058 -2.0197115
#>   2:                sp3-Atoms: 5 0.0004757940 0.06756275 -0.9039180 -1.9353310
#>   3: Total Molweight: [440, 469] 0.0015232292 0.08651942  0.7681159  2.0527086
#>   4:           H-Donors: [6, 40] 0.0015105740 0.08651942  0.8884892  1.9617057
#>   5:          Aromatic Atoms: 16 0.0015232292 0.08651942  0.7681159  2.0527086
#>  ---                                                                          
#> 280:                 H-Donors: 1 0.9552397115 0.96888599 -0.2272251 -0.5967703
#> 281: Total Molweight: [294, 312] 0.9841704124 0.98764805 -0.5088968 -0.6791230
#> 282:           Stereo Centers: 3 0.9841704124 0.98764805 -0.5088968 -0.6791230
#> 283:              Small Rings: 3 0.9822584012 0.98764805 -0.1975991 -0.5376399
#> 284:           Rings Closures: 3 0.9956051024 0.99560510 -0.1762806 -0.4527586
#>      nMoreExtreme size                                   leadingEdge
#>   1:            0   13   737754,752782,757441,1011,732517,759877,...
#>   2:            3    5                          174939,737754,752782
#>   3:            1    6        759856,715055,758247,756645,741078,740
#>   4:            2    4                   125066,758612,759155,123127
#>   5:            1    6        759856,715055,750691,765694,718781,740
#>  ---                                                                
#> 280:         9005   11            656576,758487,764040,753686,747856
#> 281:         5035    1                                        613327
#> 282:         5035    1                                        613327
#> 283:         9411   13 706995,683864,339555,226080,758664,606698,...
#> 284:         9287   10                                        706995