Analyse drug set enrichment

Analyse drug set enrichment

analyseDrugSetEnrichment(
  sets,
  stats,
  col = NULL,
  nperm = 10000,
  maxSize = 500,
  ...,
  keyColSets = NULL,
  keyColStats = NULL
)

Arguments

sets

Named list of characters: named sets containing compound identifiers (obtain drug sets by running prepareDrugSets())

stats

Named numeric vector or either a similarPerturbations or a targetingDrugs object (obtained after running rankSimilarPerturbations or predictTargetingDrugs, respectively)

col

Character: name of the column to use for statistics (only required if class of stats is either similarPerturbations or targetingDrugs)

nperm

Number of permutations to do. Minimial possible nominal p-value is about 1/nperm

maxSize

Maximal size of a gene set to test. All pathways above the threshold are excluded.

...

Arguments passed on to fgsea::fgseaSimple

minSize

Minimal size of a gene set to test. All pathways below the threshold are excluded.

scoreType

This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")

nproc

If not equal to zero sets BPPARAM to use nproc workers (default = 0).

gseaParam

GSEA parameter value, all gene-level statis are raised to the power of `gseaParam` before calculation of GSEA enrichment scores.

BPPARAM

Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting `nproc` default value `bpparam()` is used.

keyColSets

Character: column from sets to compare with column keyColStats from stats; automatically selected if NULL

keyColStats

Character: column from stats to compare with column keyColSets from sets; automatically selected if NULL

Value

Enrichment analysis based on GSEA

See also

Examples

descriptors <- loadDrugDescriptors()
#> compound_descriptors_NCI60_2D.qs not found: downloading data...
drugSets <- prepareDrugSets(descriptors)

# Analyse drug set enrichment in ranked targeting drugs for a differential
# expression profile
data("diffExprStat")
gdsc      <- loadExpressionDrugSensitivityAssociation("GDSC")
#> expressionDrugSensitivityCorGDSC7.qs not found: downloading data...
#> Loading data from expressionDrugSensitivityCorGDSC7.qs...
predicted <- predictTargetingDrugs(diffExprStat, gdsc)
#> Subsetting data based on 11396 intersecting genes (85% of the 13451 input genes)...
#> Comparing against 266 GDSC 7 compounds (983 cell lines) using 'spearman, pearson, gsea' (gene size of 150)...
#> Comparison performed in 3.21 secs

analyseDrugSetEnrichment(drugSets, predicted)
#> Matching compounds with those available in drug sets...
#> Ordering results by column 'rankProduct_rank'; to manually select column to order by, please set argument 'col'
#> Columns 'name' and 'name' were matched based on 56 common values; to manually select columns to compare, please set arguments starting with 'keyCol'
#> Performing enrichment analysis...
#>                          descriptor         pval       padj         ES
#>   1:                 Small Rings: 4 0.0002066116 0.02933884 -0.6915988
#>   2:                   sp3-Atoms: 5 0.0001192748 0.02933884 -0.9037610
#>   3:              H-Donors: [6, 40] 0.0009751341 0.06923452  0.8880866
#>   4:              Non-C/H Atoms: 11 0.0007601260 0.06923452 -0.7500000
#>   5: Polar Surface Area: [123, 141] 0.0025241902 0.14337400 -0.6509521
#>  ---                                                                  
#> 280:               Non-C/H Atoms: 7 0.9261219793 0.93600940 -0.2876703
#> 281:       Electronegative Atoms: 7 0.9261219793 0.93600940 -0.2876703
#> 282:                 Small Rings: 3 0.9297735300 0.93636767 -0.2350154
#> 283:    Total Molweight: [294, 312] 0.9818616703 0.98186167 -0.5107143
#> 284:              Stereo Centers: 3 0.9818616703 0.98186167 -0.5107143
#>             NES nMoreExtreme size                                   leadingEdge
#>   1: -1.9302423            1   15   737754,752782,757441,1011,732517,759877,...
#>   2: -1.9354953            0    5                          174939,737754,752782
#>   3:  1.9462701            1    4                   125066,758612,759155,123127
#>   4: -1.8750640            6    9 299933,330507,727989,745750,732517,759877,...
#>   5: -1.7396839           23   12  174939,737754,752782,287459,63878,732517,...
#>  ---                                                                           
#> 280: -0.6484645         8047    6                                        706995
#> 281: -0.6484645         8047    6                                        706995
#> 282: -0.6477692         8949   14 299933,706995,683864,339555,606698,226080,...
#> 283: -0.6820882         4925    1                                        613327
#> 284: -0.6820882         4925    1                                        613327