Download metadata for ENCODE knockdown experiments

Download metadata for ENCODE knockdown experiments

downloadENCODEknockdownMetadata(
  cellLine = NULL,
  gene = NULL,
  file = "ENCODEmetadata.rds"
)

Arguments

cellLine

Character: cell line

gene

Character: target gene

file

Character: RDS filepath with metadata (if file doesn't exist, it will be created)

Value

Data frame containing ENCODE knockdown experiment metadata

See also

Examples

downloadENCODEknockdownMetadata("HepG2", "EIF4G1")
#> Downloading metadata for ENCODE knockdown experiments...
#>    File accession File format File type File format type          Output type
#> 1:    ENCFF955TXI         tsv       tsv                  gene quantifications
#> 2:    ENCFF049UZV         tsv       tsv                  gene quantifications
#>    File assembly Experiment accession         Assay                   Donor(s)
#> 1:          hg19          ENCSR509LIV shRNA RNA-seq /human-donors/ENCDO000AAC/
#> 2:          hg19          ENCSR509LIV shRNA RNA-seq /human-donors/ENCDO000AAC/
#>    Biosample term id Biosample term name Biosample type Biosample organism
#> 1:       EFO:0001187               HepG2      cell line       Homo sapiens
#> 2:       EFO:0001187               HepG2      cell line       Homo sapiens
#>    Biosample treatments Biosample treatments amount
#> 1:                   NA                          NA
#> 2:                   NA                          NA
#>    Biosample treatments duration Biosample genetic modifications methods
#> 1:                            NA                                    RNAi
#> 2:                            NA                                    RNAi
#>    Biosample genetic modifications categories
#> 1:                               interference
#> 2:                               interference
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          Biosample genetic modifications targets
#> 1: {'schema_version': '14', 'aliases': [], 'organism': {'schema_version': '6', '@type': ['Organism', 'Item'], 'name': 'human', 'taxon_id': '9606', 'scientific_name': 'Homo sapiens', '@id': '/organisms/human/', 'uuid': '7745b647-ff15-4ff3-9ced-b897d4e2983c', 'status': 'released'}, 'genes': ['/genes/1981/'], '@type': ['Target', 'Item'], 'name': 'EIF4G1-human', 'label': 'EIF4G1', '@id': '/targets/EIF4G1-human/', 'title': 'EIF4G1 (Homo sapiens)', 'investigated_as': ['RNA binding protein'], 'uuid': 'cc45c7a4-2d7a-4e86-9ca0-7429983a0aab', 'status': 'released'}
#> 2: {'schema_version': '14', 'aliases': [], 'organism': {'schema_version': '6', '@type': ['Organism', 'Item'], 'name': 'human', 'taxon_id': '9606', 'scientific_name': 'Homo sapiens', '@id': '/organisms/human/', 'uuid': '7745b647-ff15-4ff3-9ced-b897d4e2983c', 'status': 'released'}, 'genes': ['/genes/1981/'], '@type': ['Target', 'Item'], 'name': 'EIF4G1-human', 'label': 'EIF4G1', '@id': '/targets/EIF4G1-human/', 'title': 'EIF4G1 (Homo sapiens)', 'investigated_as': ['RNA binding protein'], 'uuid': 'cc45c7a4-2d7a-4e86-9ca0-7429983a0aab', 'status': 'released'}
#>    Biosample genetic modifications gene targets
#> 1:                                           NA
#> 2:                                           NA
#>    Biosample genetic modifications site coordinates
#> 1:                                               NA
#> 2:                                               NA
#>    Biosample genetic modifications zygosity Experiment target
#> 1:                                       NA            EIF4G1
#> 2:                                       NA            EIF4G1
#>      Library made from Library depleted in
#> 1: polyadenylated mRNA                    
#> 2: polyadenylated mRNA                    
#>                                    Library extraction method
#> 1: Maxwell 16 LEV simpleRNA Cells Kit (Promega cat#: AS1270)
#> 2: Maxwell 16 LEV simpleRNA Cells Kit (Promega cat#: AS1270)
#>    Library lysis method Library crosslinking method Library strand specific
#> 1:                   NA                          NA                 reverse
#> 2:                   NA                          NA                 reverse
#>    Experiment date released Project RBNS protein concentration
#> 1:               2015-08-18  ENCODE                         NA
#> 2:               2015-08-18  ENCODE                         NA
#>    Library fragmentation method Library size range Biological replicate(s)
#> 1:   chemical (Illumina TruSeq)               >200                       1
#> 2:   chemical (Illumina TruSeq)               >200                       2
#>    Technical replicate(s) Read length Mapped read length Run type Paired end
#> 1:                    1_1          NA                 NA                  NA
#> 2:                    2_1          NA                 NA                  NA
#>    Paired with Index of                             Derived from    Size
#> 1:                   NA /files/ENCFF995LZN/, /files/ENCFF826ONU/ 9554701
#> 2:                   NA /files/ENCFF447GOE/, /files/ENCFF826ONU/ 9563218
#>                           Lab                           md5sum dbxrefs
#> 1: ENCODE Processing Pipeline f9b46cbff89bfe91bdb0879b81cf41e8        
#> 2: ENCODE Processing Pipeline 9f77694ffdf48801edb9c39f75f6d3b7        
#>                                                             File download URL
#> 1: https://www.encodeproject.org/files/ENCFF955TXI/@@download/ENCFF955TXI.tsv
#> 2: https://www.encodeproject.org/files/ENCFF049UZV/@@download/ENCFF049UZV.tsv
#>    Genome annotation Platform Controlled by File Status
#> 1:               V19                           released
#> 2:               V19                           released
#>                                                                                s3_uri
#> 1: s3://encode-public/2015/08/26/f02dd75d-a0c9-451f-aae8-23eb5e8a5a2f/ENCFF955TXI.tsv
#> 2: s3://encode-public/2015/08/26/d03e3680-479e-470b-9126-198e6cefe4dc/ENCFF049UZV.tsv
#>                                                                                                                                                                                            Azure URL
#> 1: https://datasetencode.blob.core.windows.net/dataset/2015/08/26/f02dd75d-a0c9-451f-aae8-23eb5e8a5a2f/ENCFF955TXI.tsv?sv=2019-10-10&si=prod&sr=c&sig=9qSQZo4ggrCNpybBExU8SypuUZV33igI11xw0P7rB3c%3D
#> 2: https://datasetencode.blob.core.windows.net/dataset/2015/08/26/d03e3680-479e-470b-9126-198e6cefe4dc/ENCFF049UZV.tsv?sv=2019-10-10&si=prod&sr=c&sig=9qSQZo4ggrCNpybBExU8SypuUZV33igI11xw0P7rB3c%3D
#>         File analysis title File analysis status
#> 1: Mixed pipelines hg19 V19             released
#> 2: Mixed pipelines hg19 V19             released
#>                         Audit WARNING Audit NOT_COMPLIANT Audit ERROR
#> 1: missing biosample characterization                              NA
#> 2: missing biosample characterization                              NA
#>        Control      Control Experiments
#> 1: ENCSR042QTH ENCFF657KMW, ENCFF726AVT
#> 2: ENCSR042QTH ENCFF657KMW, ENCFF726AVT